feat: add build phase for reading package.json
