+++ packages/eas-build-job/src/logs.ts
@@ -4,6 +4,7 @@ export enum BuildPhase {
   SPIN_UP_BUILDER = 'SPIN_UP_BUILDER',
   BUILDER_INFO = 'BUILDER_INFO',
   READ_APP_CONFIG = 'READ_APP_CONFIG',
+  READ_PACKAGE_JSON = 'READ_PACKAGE_JSON',
   SET_UP_BUILD_ENVIRONMENT = 'SET_UP_BUILD_ENVIRONMENT',
   START_BUILD = 'START_BUILD',
   INSTALL_CUSTOM_TOOLS = 'INSTALL_CUSTOM_TOOLS',
